maked corrections of -Wall -Werror -Wextra

+++ sources/handleReq/ChunkedBodyProcess.hpp
@@ -6,7 +6,7 @@
 /*   By: dsilveri <dsilveri@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/02 14:19:13 by dsilveri          #+#    #+#             */
-/*   Updated: 2023/09/11 08:40:56 by dsilveri         ###   ########.fr       */
+/*   Updated: 2023/09/23 15:01:55 by dsilveri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 class ChunkedBodyProcess: public IState
 {
 	private:
-		int _hexToDec(std::string hex);
+		size_t _hexToDec(std::string hex);
 
 	public:
 		ChunkedBodyProcess(void);
